fix upsert tasks

+++ models/tasks.go
@@ -23,6 +23,6 @@ type Tasks struct {
 	CreatedAt     time.Time      `db:"created_at" json:"created_at"`
 	UpdatedAt     time.Time      `db:"updated_at" json:"updated_at"`
 	Position      Position       `db:"position" json:"position"`
-	X             int64          `db:"x" json:"x"`
-	Y             int64          `db:"y" json:"y"`
+	X             int            `db:"x" json:"x"`
+	Y             int            `db:"y" json:"y"`
 }
